requirements.txt
nltk~=3.5
requests~=2.25.0
beautifulsoup4~=4.9.3
sklearn~=0.0
scikit-learn~=0.23.2
python-telegram-bot	~= 13.1
pony~=0.7.14